requirements.txt
streamlit==1.26.0
tensorflow==2.14.0
pillow==10.0.1
numpy==1.25.2